move ollama list retrieval into renderer
